pyproject.toml
[tool.poetry]
name = "pydantic-api-models-notion"
version = "0.0.1"
description = "A Pydantic-based library for Notion API data models"
packages = [{ include = "pydantic_api" }]
authors = ["stevieflyer"]
readme = "README.md"
repository = "https://github.com/stevieflyer/pydantic-api-models-notion"
homepage = "https://github.com/stevieflyer/pydantic-api-models-notion"
keywords = ["notion", "pydantic", "api", "data-models"]
license = "MIT"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries",
]

[tool.poetry.dependencies]
python = "^3.8"
emoji = "^2.14.0"
pydantic-api-models = { git = "ssh://git@github.com/stevieflyer/pydantic-api-models.git" }

[build-system]
requires = ["poetry-core>=1.5.0"]
build-backend = "poetry.core.masonry.api"
